requirements.txt
flask
argon2-cffi
cryptography
python-dotenv